resturture main.go
- create Command interface to return subcommands
- move each subcommand to separate file
- create flags.go to house all flags
